Create install_script_fedora.sh
This is a script which is designed to install chadwm and it's requirements on Fedora Linux.

Following the provided vanilla instructions doesn't fully lead one to a working chadwm install. For example, eww was not listed as requirement and I personally had never heard of it before. Another example being the installation of the necessary nerd fonts needed to show both text and icons (icons do not work with the tff file in vanilla instructions).

What follows is a script which is intended to both automate and inform the user as to the proper installation instructions for this distro.